app/requirements.txt
alembic==1.12.0
fastapi==0.103.2
sqlmodel~=0.0.8
sqlalchemy[asyncio]==1.4.41
asyncpg==0.28.0
psycopg2-binary==2.9.8
pydantic[email]==1.10.9
Babel~=2.12.1
tenacity==8.2.3
python-dotenv~=1.0.0
beautifulsoup4~=4.12.2
aiohttp~=3.8.4
starlette-admin==0.11.2
passlib[bcrypt]~=1.7.4
itsdangerous~=2.1.2
async-cron~=1.6.2
playwright~=1.38.0
python-jose[cryptography]==3.3.0
